generate the fallback default import block for named external imports
If an external module is traced (ExternalModule.traceExport) to
export both as default and also named imports, interop generates
the following default import block:

```js
    var foo__default = foo['default'];
```

but if that module was not external, it would generate the equivalent
of:

```js
    var foo__default = 'default' in foo ? foo['default'] : foo;
```

This happens because getInteropBlock() short-circuits when
a module is imported as a namespace, regardless of it being
also being imported with named imports.

I don't know why in the namespace-import-only case it's preferred
not to generate the fallback expression. I just left it as-is
and only took precedence when the module usage has a named import.

Added a test-case:

`tests/forms/samples/import-external-namespace-default-and-named`

and modified an existing test:

`tests/forms/samples/import-external-namespace-and-default`